+++ src/table/src/error.rs
@@ -138,13 +138,6 @@ pub enum Error {
         location: Location,
     },
 
-    #[snafu(display("Changing table option `{}` is not supported", key))]
-    UnsupportedTableOptionChange {
-        key: String,
-        #[snafu(implicit)]
-        location: Location,
-    },
-
     #[snafu(display("Table options value is not valid, key: `{}`, value: `{}`", key, value))]
     InvalidTableOptionValue { key: String, value: String },
 }
@@ -167,7 +160,6 @@ impl ErrorExt for Error {
             Error::Unsupported { .. } => StatusCode::Unsupported,
             Error::ParseTableOption { .. } => StatusCode::InvalidArguments,
             Error::MissingTimeIndexColumn { .. } => StatusCode::IllegalState,
-            Error::UnsupportedTableOptionChange { .. } => StatusCode::Unsupported,
             Error::InvalidTableOptionValue { .. } => StatusCode::InvalidArguments,
         }
     }
